refactor: extract chat generation lifecycle handlers
